Adds code refactoring for pagination
